backend/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography
defusedxml
# Django==4.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
social-auth-app-django
social-auth-core==4.4.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.2
